requirements.txt
tensorflow
opencv-python
numpy
matplotlib
scikit-learn
pydicom
pandas
seaborn
streamlit
Pillow
